Change func name scheck_starving to check_eaten. Init monitor_pid. Error msg for  pthread modified.

+++ philo_bonus/main_bonus.c
@@ -6,7 +6,7 @@
 /*   By: irychkov <irychkov@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/21 16:50:59 by irychkov          #+#    #+#             */
-/*   Updated: 2024/11/27 18:39:04 by irychkov         ###   ########.fr       */
+/*   Updated: 2024/11/27 23:09:44 by irychkov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,9 @@ static void	monitor_eating_completion(t_program_data *data)
 	exit(0);
 }
 
-static void	check_starving(t_program_data *data, pid_t *monitor_pid)
+static void	check_eaten(t_program_data *data, pid_t *monitor_pid)
 {
+	*monitor_pid = -1;
 	if (data->number_of_times_each_philosopher_must_eat != -1)
 	{
 		*monitor_pid = fork();
@@ -63,7 +64,7 @@ static int	kill_processes(t_program_data *data, pid_t monitor_pid)
 		if (waitpid(data->philos[i].pid, &status, 0) == -1)
 			exit_code = error_and_return("Failed to wait philo process\n", 1);
 		if (!WIFSIGNALED(status))
-			exit_code = error_and_return("Failed pthread in philo process\n", 1);
+			exit_code = error_and_return("Pthread error in philo process\n", 1);
 		i++;
 	}
 	return (exit_code);
@@ -75,23 +76,23 @@ static int	run_philos(t_program_data *data)
 	int		exit_code;
 	pid_t	monitor_pid;
 
-	i = 0;
+	i = -1;
 	exit_code = 0;
-	check_starving(data, &monitor_pid);
+	check_eaten(data, &monitor_pid);
 	data->start_time = get_current_time();
-	while (i < data->number_of_philosophers)
+	while (++i < data->number_of_philosophers)
 	{
 		data->philos[i].pid = fork();
 		if (data->philos[i].pid == -1)
 		{
-			kill(monitor_pid, 9);
+			if (data->number_of_times_each_philosopher_must_eat != -1)
+				kill(monitor_pid, 9);
 			while (i > 0)
 				kill(data->philos[--i].pid, 9);
 			error_and_return("Error: fork failed for philos\n", 1);
 		}
 		if (data->philos[i].pid == 0)
 			philosopher_routine(&data->philos[i]);
-		i++;
 	}
 	sem_wait(data->start);
 	exit_code = kill_processes(data, monitor_pid);
